Refactor Products to single component per file
