+++ examples/demo/src/products/index.js
@@ -1,215 +1,11 @@
-import React from 'react';
-import {
-    translate,
-    Create,
-    Datagrid,
-    DateField,
-    Edit,
-    EditButton,
-    Filter,
-    FormTab,
-    List,
-    NumberInput,
-    Pagination,
-    ReferenceInput,
-    ReferenceManyField,
-    SearchInput,
-    SelectInput,
-    TabbedForm,
-    TextField,
-    TextInput,
-    required,
-} from 'react-admin';
-import Chip from '@material-ui/core/Chip';
-import withStyles from '@material-ui/core/styles/withStyles';
-import Icon from '@material-ui/icons/Collections';
-import RichTextInput from 'ra-input-rich-text';
+import ProductIcon from '@material-ui/icons/Collections';
+import ProductList from './ProductList';
+import ProductEdit from './ProductEdit';
+import ProductCreate from './ProductCreate';
 
-import CustomerReferenceField from '../visitors/CustomerReferenceField';
-import StarRatingField from '../reviews/StarRatingField';
-import GridList from './GridList';
-import Poster from './Poster';
-
-export const ProductIcon = Icon;
-
-const quickFilterStyles = {
-    root: {
-        marginBottom: '0.7em',
-    },
+export default {
+    list: ProductList,
+    create: ProductCreate,
+    edit: ProductEdit,
+    icon: ProductIcon,
 };
-
-const QuickFilter = translate(
-    withStyles(quickFilterStyles)(({ classes, label, translate }) => (
-        <Chip className={classes.root} label={translate(label)} />
-    ))
-);
-
-export const ProductFilter = props => (
-    <Filter {...props}>
-        <SearchInput source="q" alwaysOn />
-        <ReferenceInput
-            source="category_id"
-            reference="categories"
-            sort={{ field: 'id', order: 'ASC' }}
-        >
-            <SelectInput source="name" />
-        </ReferenceInput>
-        <NumberInput source="width_gte" />
-        <NumberInput source="width_lte" />
-        <NumberInput source="height_gte" />
-        <NumberInput source="height_lte" />
-        <QuickFilter
-            label="resources.products.fields.stock_lte"
-            source="stock_lte"
-            defaultValue={10}
-        />
-    </Filter>
-);
-
-export const ProductList = props => (
-    <List
-        {...props}
-        filters={<ProductFilter />}
-        perPage={20}
-        sort={{ field: 'id', order: 'ASC' }}
-    >
-        <GridList />
-    </List>
-);
-
-const createStyles = {
-    stock: { width: '5em' },
-    price: { width: '5em' },
-    width: { width: '5em' },
-    widthFormGroup: { display: 'inline-block' },
-    height: { width: '5em' },
-    heightFormGroup: { display: 'inline-block', marginLeft: 32 },
-};
-
-export const ProductCreate = withStyles(createStyles)(
-    ({ classes, ...props }) => (
-        <Create {...props}>
-            <TabbedForm>
-                <FormTab label="resources.products.tabs.image">
-                    <TextInput
-                        autoFocus
-                        source="image"
-                        options={{ fullWidth: true }}
-                        validate={required()}
-                    />
-                    <TextInput
-                        source="thumbnail"
-                        options={{ fullWidth: true }}
-                        validate={required()}
-                    />
-                </FormTab>
-                <FormTab label="resources.products.tabs.details" path="details">
-                    <TextInput source="reference" validate={required()} />
-                    <NumberInput
-                        source="price"
-                        validate={required()}
-                        className={classes.price}
-                    />
-                    <NumberInput
-                        source="width"
-                        validate={required()}
-                        className={classes.width}
-                        formClassName={classes.widthFormGroup}
-                    />
-                    <NumberInput
-                        source="height"
-                        validate={required()}
-                        className={classes.height}
-                        formClassName={classes.heightFormGroup}
-                    />
-                    <ReferenceInput
-                        source="category_id"
-                        reference="categories"
-                        allowEmpty
-                    >
-                        <SelectInput source="name" />
-                    </ReferenceInput>
-                    <NumberInput
-                        source="stock"
-                        validate={required()}
-                        className={classes.stock}
-                    />
-                </FormTab>
-                <FormTab
-                    label="resources.products.tabs.description"
-                    path="description"
-                >
-                    <RichTextInput source="description" addLabel={false} />
-                </FormTab>
-            </TabbedForm>
-        </Create>
-    )
-);
-
-const ProductTitle = ({ record }) => <span>Poster #{record.reference}</span>;
-
-const editStyles = {
-    ...createStyles,
-    comment: {
-        maxWidth: '20em',
-        overflow: 'hidden',
-        textOverflow: 'ellipsis',
-        whiteSpace: 'nowrap',
-    },
-};
-
-export const ProductEdit = withStyles(editStyles)(({ classes, ...props }) => (
-    <Edit {...props} title={<ProductTitle />}>
-        <TabbedForm>
-            <FormTab label="resources.products.tabs.image">
-                <Poster />
-                <TextInput source="image" options={{ fullWidth: true }} />
-                <TextInput source="thumbnail" options={{ fullWidth: true }} />
-            </FormTab>
-            <FormTab label="resources.products.tabs.details" path="details">
-                <TextInput source="reference" />
-                <NumberInput source="price" className={classes.price} />
-                <NumberInput
-                    source="width"
-                    className={classes.width}
-                    formClassName={classes.widthFormGroup}
-                />
-                <NumberInput
-                    source="height"
-                    className={classes.height}
-                    formClassName={classes.heightFormGroup}
-                />
-                <ReferenceInput source="category_id" reference="categories">
-                    <SelectInput source="name" />
-                </ReferenceInput>
-                <NumberInput source="stock" className={classes.stock} />
-            </FormTab>
-            <FormTab
-                label="resources.products.tabs.description"
-                path="description"
-            >
-                <RichTextInput source="description" addLabel={false} />
-            </FormTab>
-            <FormTab label="resources.products.tabs.reviews" path="reviews">
-                <ReferenceManyField
-                    reference="reviews"
-                    target="product_id"
-                    addLabel={false}
-                    pagination={<Pagination />}
-                >
-                    <Datagrid>
-                        <DateField source="date" />
-                        <CustomerReferenceField />
-                        <StarRatingField />
-                        <TextField
-                            source="comment"
-                            cellClassName={classes.comment}
-                        />
-                        <TextField source="status" />
-                        <EditButton />
-                    </Datagrid>
-                </ReferenceManyField>
-            </FormTab>
-        </TabbedForm>
-    </Edit>
-));
